added long full 40 hour test. refactored testing
