+++ coco/test/deploy2deter/run_tests/main.go
@@ -29,356 +29,17 @@
 package main
 
 import (
-	"bytes"
-	"encoding/json"
 	"errors"
 	"fmt"
-	"io"
-	"io/ioutil"
 	"log"
 	"math"
-	"net/http"
 	"os"
 	"os/exec"
-	"runtime"
-	"strconv"
 	"time"
-
-	"github.com/PuerkitoBio/goquery"
-	"github.com/pkg/browser"
-	"golang.org/x/net/websocket"
 )
 
-type RunStats struct {
-	NHosts int
-	Depth  int
-
-	BF int
-
-	MinTime float64
-	MaxTime float64
-	AvgTime float64
-	StdDev  float64
-
-	SysTime  float64
-	UserTime float64
-
-	Rate float64
-}
-
-func (s RunStats) CSVHeader() []byte {
-	var buf bytes.Buffer
-	buf.WriteString("hosts, depth, bf, min, max, avg, stddev, systime, usertime, rate\n")
-	return buf.Bytes()
-}
-func (s RunStats) CSV() []byte {
-	var buf bytes.Buffer
-	fmt.Fprintf(&buf, "%d, %d, %d, %f, %f, %f, %f, %f, %f, %f\n",
-		s.NHosts,
-		s.Depth,
-		s.BF,
-		s.MinTime,
-		s.MaxTime,
-		s.AvgTime,
-		s.StdDev,
-		s.SysTime,
-		s.UserTime,
-		s.Rate)
-	return buf.Bytes()
-}
-
-func RunStatsAvg(rs []RunStats) RunStats {
-	if len(rs) == 0 {
-		return RunStats{}
-	}
-	r := RunStats{}
-	r.NHosts = rs[0].NHosts
-	r.Depth = rs[0].Depth
-	r.BF = rs[0].BF
-
-	for _, a := range rs {
-		r.MinTime += a.MinTime
-		r.MaxTime += a.MaxTime
-		r.AvgTime += a.AvgTime
-		r.StdDev += a.StdDev
-		r.SysTime += a.SysTime
-		r.UserTime += a.UserTime
-		r.Rate += a.Rate
-	}
-	l := float64(len(rs))
-	r.MinTime /= l
-	r.MaxTime /= l
-	r.AvgTime /= l
-	r.StdDev /= l
-	r.SysTime /= l
-	r.UserTime /= l
-	r.Rate /= l
-	return r
-}
-
-/*
-{
-	"eapp":"time",
-	"ehost":"10.255.0.13:2000",
-	"elevel":"info",
-	"emsg":"root round",
-	"etime":"2015-02-27T09:50:45-08:00",
-	"file":"server.go:195",
-	"round":59,
-	"time":893709029,
-	"type":"root_round"
-}
-*/
-var view bool
-var debug string = "-debug=false"
-
-func SetDebug(b bool) {
-	if b {
-		debug = "-debug=true"
-	} else {
-		debug = "-debug=false"
-	}
-}
-
-type StatsEntry struct {
-	App     string  `json:"eapp"`
-	Host    string  `json:"ehost"`
-	Level   string  `json:"elevel"`
-	Msg     string  `json:"emsg"`
-	MsgTime string  `json:"etime"`
-	File    string  `json:"file"`
-	Round   int     `json:"round"`
-	Time    float64 `json:"time"`
-	Type    string  `json:"type"`
-}
-
-type SysStats struct {
-	File     string  `json:"file"`
-	Type     string  `json:"type"`
-	SysTime  float64 `json:"systime"`
-	UserTime float64 `json:"usertime"`
-}
-
-type ClientMsgStats struct {
-	File        string    `json:"file"`
-	Type        string    `json:"type"`
-	Buckets     []float64 `json:"buck,omitempty"`
-	RoundsAfter []float64 `json:"roundsAfter,omitempty"`
-}
-
-type ExpVar struct {
-	Cmdline  []string         `json:"cmdline"`
-	Memstats runtime.MemStats `json:"memstats"`
-}
-
-func Memstats(server string) (*ExpVar, error) {
-	url := "localhost:8080/d/" + server + "/debug/vars"
-	resp, err := http.Get(url)
-	if err != nil {
-		return nil, err
-	}
-	b, err := ioutil.ReadAll(resp.Body)
-	resp.Body.Close()
-	if err != nil {
-		return nil, err
-	}
-	var evar ExpVar
-	err = json.Unmarshal(b, &evar)
-	if err != nil {
-		log.Println("failed to unmarshal expvar:", string(b))
-		return nil, err
-	}
-	return &evar, nil
-}
-
-func MonitorMemStats(server string, poll int, done chan struct{}, stats *[]*ExpVar) {
-	go func() {
-		ticker := time.NewTicker(time.Duration(poll) * time.Millisecond)
-		for {
-			select {
-			case <-ticker.C:
-				evar, err := Memstats(server)
-				if err != nil {
-					continue
-				}
-				*stats = append(*stats, evar)
-			case <-done:
-				return
-			}
-		}
-	}()
-}
-
-// Monitor: monitors log aggregates results into RunStats
-func Monitor(bf int) RunStats {
-	log.Println("MONITORING")
-	defer fmt.Println("DONE MONITORING")
-retry_dial:
-	ws, err := websocket.Dial("ws://localhost:8080/log", "", "http://localhost/")
-	if err != nil {
-		time.Sleep(1 * time.Second)
-		goto retry_dial
-	}
-retry:
-	// Get HTML of webpage for data (NHosts, Depth, ...)
-	doc, err := goquery.NewDocument("http://localhost:8080/")
-	if err != nil {
-		log.Println("unable to get log data: retrying:", err)
-		time.Sleep(10 * time.Second)
-		goto retry
-	}
-	if view {
-		browser.OpenURL("http://localhost:8080/")
-	}
-	nhosts := doc.Find("#numhosts").First().Text()
-	log.Println("hosts:", nhosts)
-	depth := doc.Find("#depth").First().Text()
-	log.Println("depth:", depth)
-	nh, err := strconv.Atoi(nhosts)
-	if err != nil {
-		log.Fatal("unable to convert hosts to be a number:", nhosts)
-	}
-	d, err := strconv.Atoi(depth)
-	if err != nil {
-		log.Fatal("unable to convert depth to be a number:", depth)
-	}
-	client_done := false
-	root_done := false
-	var rs RunStats
-	rs.NHosts = nh
-	rs.Depth = d
-	rs.BF = bf
-
-	var M, S float64
-	k := float64(1)
-	first := true
-	for {
-		var data []byte
-		err := websocket.Message.Receive(ws, &data)
-		if err != nil {
-			// if it is an eof error than stop reading
-			if err == io.EOF {
-				log.Println("websocket terminated before emitting EOF or terminating string")
-				break
-			}
-			continue
-		}
-		if bytes.Contains(data, []byte("EOF")) || bytes.Contains(data, []byte("terminating")) {
-			log.Printf(
-				"EOF/terminating Detected: need forkexec to report and clients: %b %b",
-				root_done, client_done)
-		}
-		if bytes.Contains(data, []byte("root_round")) {
-			if client_done || root_done {
-				// ignore after we have received our first EOF
-				continue
-			}
-			var entry StatsEntry
-			err := json.Unmarshal(data, &entry)
-			if err != nil {
-				log.Fatal("json unmarshalled improperly:", err)
-			}
-			log.Println("root_round:", entry)
-			if first {
-				first = false
-				rs.MinTime = entry.Time
-				rs.MaxTime = entry.Time
-			}
-			if entry.Time < rs.MinTime {
-				rs.MinTime = entry.Time
-			} else if entry.Time > rs.MaxTime {
-				rs.MaxTime = entry.Time
-			}
-
-			rs.AvgTime = ((rs.AvgTime * (k - 1)) + entry.Time) / k
-
-			var tM = M
-			M += (entry.Time - tM) / k
-			S += (entry.Time - tM) * (entry.Time - M)
-			k++
-			rs.StdDev = math.Sqrt(S / (k - 1))
-		} else if bytes.Contains(data, []byte("forkexec")) {
-			if root_done {
-				continue
-			}
-			var ss SysStats
-			err := json.Unmarshal(data, &ss)
-			if err != nil {
-				log.Fatal("unable to unmarshal forkexec:", ss)
-			}
-			rs.SysTime = ss.SysTime
-			rs.UserTime = ss.UserTime
-			log.Println("FORKEXEC:", ss)
-			if client_done {
-				break
-			}
-			root_done = true
-		} else if bytes.Contains(data, []byte("client_msg_stats")) {
-			if client_done {
-				continue
-			}
-			var cms ClientMsgStats
-			err := json.Unmarshal(data, &cms)
-			if err != nil {
-				log.Fatal("unable to unmarshal client_msg_stats:", string(data))
-			}
-			// what do I want to keep out of the Client Message States
-			// cms.Buckets stores how many were processed at time T
-			// cms.RoundsAfter stores how many rounds delayed it was
-			//
-			// get the average delay (roundsAfter), max and min
-			// get the total number of messages timestamped
-			// get the average number of messages timestamped per second?
-			avg, _, _, _ := ArrStats(cms.Buckets)
-			// get the observed rate of processed messages
-			// avg is how many messages per second, we want how many milliseconds between messages
-			observed := avg / 1000 // set avg to messages per milliseconds
-			observed = 1 / observed
-			rs.Rate = observed
-			if root_done {
-				break
-			}
-			client_done = true
-		}
-	}
-	return rs
-}
-
-func ArrStats(stream []float64) (avg float64, min float64, max float64, stddev float64) {
-	// truncate trailing 0s
-	i := len(stream) - 1
-	for ; i >= 0; i-- {
-		if math.Abs(stream[i]) > 0.01 {
-			break
-		}
-	}
-	stream = stream[:i+1]
-
-	k := float64(1)
-	first := true
-	var M, S float64
-	for _, e := range stream {
-		if first {
-			first = false
-			min = e
-			max = e
-		}
-		if e < min {
-			min = e
-		} else if max < e {
-			max = e
-		}
-		avg = ((avg * (k - 1)) + e) / k
-		var tM = M
-		M += (e - tM) / k
-		S += (e - tM) * (e - M)
-		k++
-		stddev = math.Sqrt(S / (k - 1))
-	}
-	return avg, min, max, stddev
-}
-
 type T struct {
+	nmachs   int
 	hpn      int
 	bf       int
 	rate     int
@@ -386,22 +47,28 @@ type T struct {
 	failures int
 }
 
-func isZero(f float64) bool {
-	return math.Abs(f) < 0.0000001
-}
+var DefaultMachs int = 32
+
+// time-per-round * DefaultRounds = 10 * 20 = 3.3 minutes now
+// this leaves us with 7 minutes for test setup and tear-down
+var DefaultRounds int = 20
+
+var view bool
+var debug string = "-debug=false"
 
 // hpn, bf, nmsgsG
 func RunTest(t T) (RunStats, error) {
 	// add timeout for 10 minutes?
 	done := make(chan struct{})
 	var rs RunStats
+	nmachs := fmt.Sprintf("-nmachs=%d", t.nmachs)
 	hpn := fmt.Sprintf("-hpn=%d", t.hpn)
 	nmsgs := fmt.Sprintf("-nmsgs=%d", -1)
 	bf := fmt.Sprintf("-bf=%d", t.bf)
 	rate := fmt.Sprintf("-rate=%d", t.rate)
 	rounds := fmt.Sprintf("-rounds=%d", t.rounds)
 	failures := fmt.Sprintf("-failures=%d", t.failures)
-	cmd := exec.Command("./deploy2deter", hpn, nmsgs, bf, rate, rounds, debug, failures)
+	cmd := exec.Command("./deploy2deter", nmachs, hpn, nmsgs, bf, rate, rounds, debug, failures)
 	log.Println("RUNNING TEST:", cmd.Args)
 	log.Println("FAILURES PERCENT:", t.failures)
 	cmd.Stdout = os.Stdout
@@ -432,17 +99,6 @@ func RunTest(t T) (RunStats, error) {
 	}
 
 	return rs, nil
-}
-
-func MkTestDir() {
-	err := os.MkdirAll("test_data/", 0777)
-	if err != nil {
-		log.Fatal("failed to make test directory")
-	}
-}
-
-func TestFile(name string) string {
-	return "test_data/" + name
 }
 
 // RunTests runs the given tests and puts the output into the
@@ -499,59 +155,77 @@ func RunTests(name string, ts []T) {
 // high and low specify how many milliseconds between messages
 func RateLoadTest(hpn, bf int) []T {
 	return []T{
-		{hpn, bf, 5000, DefaultRounds, 0}, // never send a message
-		{hpn, bf, 5000, DefaultRounds, 0}, // one per round
-		{hpn, bf, 500, DefaultRounds, 0},  // 10 per round
-		{hpn, bf, 50, DefaultRounds, 0},   // 100 per round
-		{hpn, bf, 30, DefaultRounds, 0},   // 1000 per round
+		{DefaultMachs, hpn, bf, 5000, DefaultRounds, 0}, // never send a message
+		{DefaultMachs, hpn, bf, 5000, DefaultRounds, 0}, // one per round
+		{DefaultMachs, hpn, bf, 500, DefaultRounds, 0},  // 10 per round
+		{DefaultMachs, hpn, bf, 50, DefaultRounds, 0},   // 100 per round
+		{DefaultMachs, hpn, bf, 30, DefaultRounds, 0},   // 1000 per round
 	}
 }
 
 func DepthTest(hpn, low, high, step int) []T {
 	ts := make([]T, 0)
 	for bf := low; bf <= high; bf += step {
-		ts = append(ts, T{hpn, bf, 10, DefaultRounds, 0})
+		ts = append(ts, T{DefaultMachs, hpn, bf, 10, DefaultRounds, 0})
 	}
 	return ts
 }
 
-func DepthTestFixed(hpn) []T {
+func DepthTestFixed(hpn int) []T {
 	return []T{
-		{hpn, 1, 30, DefaultRounds, 0},
-		{hpn, 2, 30, DefaultRounds, 0},
-		{hpn, 4, 30, DefaultRounds, 0},
-		{hpn, 8, 30, DefaultRounds, 0},
-		{hpn, 16, 30, DefaultRounds, 0},
-		{hpn, 32, 30, DefaultRounds, 0},
-		{hpn, 64, 30, DefaultRounds, 0},
-		{hpn, 128, 30, DefaultRounds, 0},
-		{hpn, 256, 30, DefaultRounds, 0},
-		{hpn, 512, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 1, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 2, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 4, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 8, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 16, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 32, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 64, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 128, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 256, 30, DefaultRounds, 0},
+		{DefaultMachs, hpn, 512, 30, DefaultRounds, 0},
 	}
 }
 
 func ScaleTest(bf, low, high, mult int) []T {
 	ts := make([]T, 0)
 	for hpn := low; hpn <= high; hpn *= mult {
-		ts = append(ts, T{hpn, bf, 10, DefaultRounds, 0})
+		ts = append(ts, T{DefaultMachs, hpn, bf, 10, DefaultRounds, 0})
 	}
 	return ts
 }
 
-var DefaultRounds int = 100
-
 // hpn=1, bf=2, rate=5000, failures=20
-var TestT = []T{
-	{1, 2, 5000, 5, 10},
-	{10, 2, 5000, 5, 5},
-	{1, 2, 5000, 5, 0},
+var FailureTests = []T{
+	{DefaultMachs, 1, 2, 5000, 5, 10},
+	{DefaultMachs, 10, 2, 5000, 5, 5},
+	{DefaultMachs, 1, 2, 5000, 5, 0},
+}
+
+func FullTests() []T {
+	var nmachs = []int{1, 16, 32}
+	var hpns = []int{1, 16, 32, 128}
+	var bfs = []int{2, 4, 8, 16, 128}
+	var rates = []int{5000, 500, 100, 30}
+	failures := 0
+
+	var tests []T
+	for _, nmach := range nmachs {
+		for _, hpn := range hpns {
+			for _, bf := range bfs {
+				for _, rate := range rates {
+					tests = append(tests, T{nmach, hpn, bf, rate, DefaultRounds, failures})
+				}
+			}
+		}
+	}
+
+	return tests
 }
 
 func main() {
+	SetDebug(false)
 	// view = true
 	os.Chdir("..")
-	SetDebug(true)
-	DefaultRounds = 5 // NOTE: changed from 10
 
 	MkTestDir()
 
@@ -561,20 +235,43 @@ func main() {
 	}
 	// test the testing framework
 
-	t := TestT
-	RunTests("test.csv", t)
-	/*
-		t := ScaleTest(10, 1, 100, 2)
-		RunTests("scale_test.csv", t)
-		// how does the branching factor effect speed
-		t = DepthTest(100, 2, 100, 1)
-		RunTests("depth_test.csv", t)
+	// DefaultRounds = 5
+	// t := FailureTests
+	// RunTests("failure_test.csv", t)
 
-		// load test the client
-		t = RateLoadTest(40, 10)
-		RunTests("load_rate_test_bf10.csv", t)
-		t = RateLoadTest(40, 50)
-		RunTests("load_rate_test_bf50.csv", t)
-	*/
+	t := ScaleTest(10, 1, 100, 2)
+	RunTests("scale_test.csv", t)
+	// how does the branching factor effect speed
+	t = DepthTestFixed(100)
+	RunTests("depth_test.csv", t)
 
+	// load test the client
+	t = RateLoadTest(40, 10)
+	RunTests("load_rate_test_bf10.csv", t)
+	t = RateLoadTest(40, 50)
+	RunTests("load_rate_test_bf50.csv", t)
+
+}
+
+func MkTestDir() {
+	err := os.MkdirAll("test_data/", 0777)
+	if err != nil {
+		log.Fatal("failed to make test directory")
+	}
+}
+
+func TestFile(name string) string {
+	return "test_data/" + name
+}
+
+func SetDebug(b bool) {
+	if b {
+		debug = "-debug=true"
+	} else {
+		debug = "-debug=false"
+	}
+}
+
+func isZero(f float64) bool {
+	return math.Abs(f) < 0.0000001
 }
